Changes to have a reusable function for C C++ interop optimisation log data
