+++ .gitignore
@@ -36,3 +36,5 @@ tests/Release/*
 *.sdf
 *.idb
 *.db
+.vscode/
+*.opendb
